requirements.txt
flask==2.3.2
Flask-Cors==4.0.0
google-generativeai==0.8.5
python-dotenv==1.0.0
pandas==2.1.1
xlrd==2.0.1